Fixed Major Expense Tracking issues
Issue -> The page that was created was followed by being deleted instead of the one in the original form.
The folder in google drive was created with the ID of the entry of the form Database entry instead of the new database entry/page ID, this has been adressed and fixed.